backend/requirements.txt
fastapi
sqlmodel
sqlite3
uvicorn
python-jose[cryptography]
passlib[argon2]
python-dotenv
requests
beautifulsoup4
gradio-client
Pillow